+++ src/indexer/subsquid_client.rs
@@ -28,7 +28,7 @@ struct CommitmentsData {
 #[serde(rename_all = "camelCase")]
 pub struct Commitment {
     pub hash: String,
-    pub tree_number: i32,
+    pub tree_number: u16,
     pub tree_position: i32,
     pub block_number: String,
     pub commitment_type: String,
